requirements.txt
alembic==1.15.2
bcrypt==4.3.0
blinker==1.9.0
cachelib==0.13.0
cffi==1.17.1
click==8.1.8
cmdstanpy==1.2.5
colorama==0.4.6
contourpy==1.3.2
cryptography==44.0.3
cx_Oracle==8.3.0
cycler==0.12.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.58.2
greenlet==3.2.0
holidays==0.74
importlib_resources==6.5.2
itsdangerous==2.2.0
japanize-matplotlib==1.1.3
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==2.2.6
oracledb==3.1.0
packaging==25.0
pandas==2.3.0
pillow==11.2.1
prophet==1.1.7
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
SQLAlchemy==2.0.40
stanio==0.5.1
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
Werkzeug==3.1.3
WTForms==3.2.1
